+++ src/Logic/Discount.hs
@@ -9,13 +9,13 @@ import qualified Data.HashMap.Strict as HM
 import Data.Maybe (fromMaybe)
 
 
-calculateTotalCost :: [(Int, Int)] -> [Product] -> Float
+calculateTotalCost :: [(Int, Int)] -> [Product] -> Double
 calculateTotalCost items products =
-  let productMap = HM.fromList [(productID p, price p) | p <- products]
+  let productMap = HM.fromList [(productID p, realToFrac (price p)) | p <- products]
   in sum [fromMaybe 0.0 (HM.lookup pid productMap) * fromIntegral qty
          | (pid, qty) <- items, qty > 0]
 
-applyComboDiscount :: [Category] -> [(Int, Int)] -> [Product] -> Float -> Float
+applyComboDiscount :: [Category] -> [(Int, Int)] -> [Product] -> Double -> Double
 applyComboDiscount cats items products total =
   let productMap = HM.fromList [(productID p, p) | p <- products]
       hasGuitar = any (\(pid,qty) -> qty > 0 && maybe False (\p -> isGuitar p cats) (HM.lookup pid productMap)) items
@@ -23,18 +23,18 @@ applyComboDiscount cats items products total =
       hasCable  = any (\(pid,qty) -> qty > 0 && maybe False (\p -> isCable p cats) (HM.lookup pid productMap)) items
   in if hasGuitar && hasAmp && hasCable then total * 0.95 else total
 
-applyStringsDiscount :: [Category] -> [(Int, Int)] -> [Product] -> Float -> Float
+applyStringsDiscount :: [Category] -> [(Int, Int)] -> [Product] -> Double -> Double
 applyStringsDiscount cats items products total =
   let productMap = HM.fromList [(productID p, p) | p <- products]
       hasGuitar = any (\(pid,qty) -> qty > 0 && maybe False (\p -> isGuitar p cats) (HM.lookup pid productMap)) items
-      stringsCost = sum [price p * fromIntegral qty
+      stringsCost = sum [realToFrac (price p) * fromIntegral qty
                         | (pid, qty) <- items, qty > 0
                         , Just p <- [HM.lookup pid productMap]
                         , isStrings p cats]
       discount = if hasGuitar && stringsCost > 0 then stringsCost * 0.10 else 0
   in total - discount
 
-calculateFinalCost :: [Category] -> [(Int, Int)] -> [Product] -> Float -> Float
+calculateFinalCost :: [Category] -> [(Int, Int)] -> [Product] -> Double -> Double
 calculateFinalCost cats items products total =
   let afterCombo = applyComboDiscount cats items products total
   in applyStringsDiscount cats items products afterCombo
